+++ src/apis/auth/queries.ts
@@ -1,12 +1,11 @@
-import { useQuery } from '@tanstack/react-query';
+import { useSuspenseQuery } from '@tanstack/react-query';
 
 import { authApi } from '.';
 import { queryKeys } from './keys';
 
 export const useGetLogin = (code: string, provider: string, redirectUri: string) => {
-  return useQuery({
+  return useSuspenseQuery({
     queryKey: queryKeys.getLogin(code),
     queryFn: () => authApi.getLogin(code, provider, redirectUri),
-    enabled: !!code,
   });
 };
