+++ src/app/(sub)/join/page.tsx
@@ -1,5 +1,6 @@
 import { queryKeys } from '@/apis/question';
 import { questionApi } from '@/apis/question';
+import { SSRSafeSuspense } from '@/components/Async';
 import HydrationProvider from '@/providers/HydrationProvider';
 
 import AnimatePresenceLayout from './components/AnimatePresenceLayout';
@@ -9,29 +10,31 @@ import JoinFunnel from './funnels/JoinFunnel';
 
 export default function Join() {
   return (
-    <FunnelProvider>
-      <Join1Provider>
-        <AnimatePresenceLayout>
-          <HydrationProvider
-            queries={[
-              {
-                queryKey: queryKeys.getValue('life'),
-                queryFn: () => questionApi.getValue('life'),
-              },
-              {
-                queryKey: queryKeys.getValue('job'),
-                queryFn: () => questionApi.getValue('job'),
-              },
-              {
-                queryKey: queryKeys.getValue('love'),
-                queryFn: () => questionApi.getValue('love'),
-              },
-            ]}
-          >
-            <JoinFunnel />
-          </HydrationProvider>
-        </AnimatePresenceLayout>
-      </Join1Provider>
-    </FunnelProvider>
+    <SSRSafeSuspense>
+      <FunnelProvider>
+        <Join1Provider>
+          <AnimatePresenceLayout>
+            <HydrationProvider
+              queries={[
+                {
+                  queryKey: queryKeys.getValue('life'),
+                  queryFn: () => questionApi.getValue('life'),
+                },
+                {
+                  queryKey: queryKeys.getValue('job'),
+                  queryFn: () => questionApi.getValue('job'),
+                },
+                {
+                  queryKey: queryKeys.getValue('love'),
+                  queryFn: () => questionApi.getValue('love'),
+                },
+              ]}
+            >
+              <JoinFunnel />
+            </HydrationProvider>
+          </AnimatePresenceLayout>
+        </Join1Provider>
+      </FunnelProvider>
+    </SSRSafeSuspense>
   );
 }
