--- src/app/(sub)/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return <div className="z-modal flex size-full items-center justify-center">Loading ...</div>;
-}
