+++ src/components/Async/index.ts
@@ -1,0 +1,1 @@
+export { default as SSRSafeSuspense } from './SSRSafeSuspense';
