+++ src/components/Loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return <div className="flex size-full items-center justify-center">Loading ...</div>;
+}
